flashholdDevops/requirement.txt
appnope==0.1.0
atomicwrites==1.3.0
attrs==19.2.0
autopep8==1.4.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.7.0.1
boto3==1.9.145
botocore==1.12.145
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
configparser==4.0.2
contextlib2==0.6.0.post1
decorator==4.3.2
distribute==0.7.3
Django==1.11.20
django-auth-ldap==1.7.0
django-guardian==1.5.0
django-mama-cas==2.4.0
djangorestframework==3.9.4
dnspython==1.16.0
docker==4.1.0
docutils==0.14
enum34==1.1.6
Flask==1.1.1
funcsigs==1.0.2
futures==3.3.0
gitdb2==2.0.5
GitPython==2.1.11
google-auth==1.6.3
grpcio==1.20.1
grpcio-tools==1.20.1
idna==2.8
importlib-metadata==0.23
ipaddress==1.0.22
ipython==5.8.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
kubernetes==9.0.0
Markdown==3.0.1
MarkupSafe==1.1.1
more-itertools==5.0.0
MySQL-python==1.2.5
oauthlib==3.0.1
packaging==19.2
pathlib2==2.3.3
pexpect==4.6.0
pickleshare==0.7.5
pip==19.2.3
pluggy==0.13.0
prompt-toolkit==1.0.15
protobuf==3.7.1
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
Pygments==2.3.1
pyhelm==0.1.1
pyparsing==2.4.2
pytest==4.6.5
python-dateutil==2.8.0
python-gitlab==1.8.0
python-ldap==3.2.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
s3transfer==0.2.0
scandir==1.10.0
setuptools==41.2.0
shortuuid==0.5.0
simplegeneric==0.8.1
simplejson==3.16.0
six==1.12.0
smmap2==2.0.5
supermutes==0.2.5
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.16.0
wheel==0.33.6
zipp==0.6.0
